backend/requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.6
coverage==7.3.0
dill==0.3.7
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.5.0
mypy-extensions==1.0.0
orjson==3.9.4
packaging==23.1
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pylint==2.17.4
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
